% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{ndl_harvest}
\alias{ndl_harvest}
\title{NDL Harvest API}
\usage{
ndl_harvest(params = NULL, ...)
}
\arguments{
\item{params}{A named list of parameters to be passed to the API.}

\item{...}{Additional parameters to be passed to the API.}
}
\description{
NDL Harvest API
}
\examples{
\dontrun{
ndl_harvest(params = list(verb = "GetRecord",
                          identifier = "oai:ndlsearch.ndl.go.jp:R100000002-I032561201",
                          metadataPrefix = "dcndl",
                          set = "iss-ndl-opac"))
}
}
