% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ndl_search}
\alias{ndl_search}
\title{NDL Search API}
\usage{
ndl_search(params = NULL, end_point = "OpenSearch", ...)
}
\arguments{
\item{params}{A named list of parameters to be passed to the API.}

\item{end_point}{A character string specifying the end point of the API.}

\item{...}{Additional parameters to be passed to the API.}
}
\description{
NDL Search API
}
\examples{
\dontrun{
ndl_search(creator = "\u590f\u76ee")
}
}
